Add mockapi package
